requirements.txt
amqp==5.1.1
aniso8601==9.0.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.2.35
Django==4.0
django-admin==2.0.2
django-celery-beat==2.5.0
django-excel-response2==3.0.5
django-rest-swagger==2.2.0
django-six==1.0.5
django-timezone-field==5.0
djangorestframework==3.14.0
drf-spectacular==0.26.2
excel-base==1.0.4
Flask==2.2.3
flask-restx==1.1.0
idna==3.4
inflection==0.5.1
isoweek==1.3.3
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kombu==5.2.4
MarkupSafe==2.1.2
meson==1.1.0
numpy==1.24.3
openapi-codec==1.3.2
prompt-toolkit==3.0.38
protobuf==4.21.9
pyrsistent==0.19.3
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==4.5.4
requests==2.29.0
scikit-learn==1.2.2
scipy==1.10.1
screen==1.0.1
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.1.0
TimeConvert==3.0.13
tzdata==2023.3
tzlocal==4.3
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
xlwt==1.3.0
